refac: not claimed tokens
