+++ src/controllers/user.ts
@@ -1,7 +1,7 @@
-import { ErrorResponseModel, RankRequestResult, RequestResult, Token, User } from '../types'
+import { ErrorResponseModel, RankRequestResult, RequestResult, User } from '../types'
 import { Controller, Get, Route, Security, Response } from 'tsoa'
 import { getRankService } from '../services/rank'
-import { getDatabaseUsers, getDatabaseUserById, getNonClaimedTokensByUser } from '../services/user'
+import { getDatabaseUsers, getDatabaseUserById } from '../services/user'
 
 @Route('/user')
 export class UserController extends Controller {
@@ -31,21 +31,6 @@ export class UserController extends Controller {
     try {
       const user = await getDatabaseUserById(userId)
       return user
-    } catch (error) {
-      console.log(error)
-    }
-  }
-
-  @Response<ErrorResponseModel>('404', 'Not Found', {
-    statusCode: 404,
-    message: 'Usuário não encontrado'
-  })
-  @Security("api_key")
-  @Get('{userId}/notClaimed')
-  public async getUserNotClaimedTokens(userId: string): Promise<string[]> {
-    try {
-      const tokens = await getNonClaimedTokensByUser(userId)
-      return tokens
     } catch (error) {
       console.log(error)
     }
